Add / extend Cap'n Proto interfaces for key drivers.
Note that while the outgoing-HTTP and email drivers may be implemented as pseudo-drivers, I think that the web publishing driver may have no such luck, as it needs to store significant amounts of data and we will want to charge that against the user's quota. Plus we don't have a great place to store the data if not in a grain.